services/document-parser-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiokafka==0.11.0
pymupdf==1.25.1
pytesseract==0.3.13
Pillow==11.0.0
minio==7.2.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
tenacity==9.0.0
